DON-310 - make suggested amounts 'sticky':
* add logic to remember suggestions per device and move it to DonationService
* remove logic to pass amounts from server to browser state
* update the point at which we tell GA about chosen amount, so it works with Stripe mode too
